requirements.txt
bcrypt==3.2.2
black==22.3.0
blinker==1.4
cffi==1.15.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
pymongo==4.1.1
python-dotenv==0.20.0
python-http-client==3.3.7
sendgrid==6.9.7
starkbank-ecdsa==2.0.3
tomli==2.0.1
Werkzeug==2.1.2
WTForms==3.0.1
